fix(cpu-docker,-constraints): fix issues to successfully build for our use case
Update the dockerfile to use root user (user `serving` wasn't working with letting the scripts save files). Force sockeye to use json-input. Update the code to pass the entire input dictionary to sockeye inference, instead of only the single text string (this is because constraints are in the rest of the json input, not in just the text). But warning, this could cause crashes that we aren't handling right now if there are keys in the input dict from the user that sockeye isn't recognizing).